Create tictac
This lab is introducing the importance of defining a function and its role. This code was repetative but it was needed to keep implement the logic of the game's rules going until the game was finished. I learned how to inspect certain parts of this game by using the React DevTools, which is alot more useful then trying to run and play the game every time.